Implemented the unit tests
It was a rabbit hole and half to get it working though.
It now properly mocks the MapManager and JimuLayerViews.
State updates and renders are properly awaited in await act() statements.

And some small fixes:
Added a unique key to each option in the select.
Added a label to the translation system.
Added the Experience Builder version I'm using.
Removed my previous attempt at awaiting renders during tests.